consolidate tests loading into cache loader
